further reorganizing of event scripts
